Fix breaking change in _miscop_get_text (#20)
The committed fix for issue #9 d685972 broke _miscop_get_text by adding an extra integer argument to the format string; this affects all gadgets with text attributed. This was also merged into main.
This commit removes the extra argument and also flips the component ID arg to a signed int.