Clean up more lint in toolbox __init__.py. (#62)
* Rework the fakeswi to use class methods.

The fakeswi interface is a little more flexible and can patch multiple
modules if it uses a class instead of the functions being bare inside
the fakeswi module.

This might not be quite so useful in general, but it seemed more
flexible to me.

* Remove redundant commented out mocking code from fakeswi.

The fakeswi module had commented out mocking code present, which has
been duplicated in saveas. In the future it might be moved to the
fakeswi module, but right now it's not necessary.

* Allow runtest.sh to take parameters.

runtest.sh now allows parameters to be passed, so we can run
`runtest.sh -v` to get more useful output.

* Clean up more lint in toolbox __init__.py.

This change reduces the lint errors significantly, reducing it to
just the base imports. Instead of using the wildcard imports, we
now import only the classes, methods and functions that we use.

This might cause some clients to break as they assume that the top
level module contains constants, but it is far cleaner. It might be
that instead of wildcard we need to specify everything, or
alternatively use __all__ in the imported modues, instead of this.

* Restore back to using 'import *' with warnings for flake8.

The 'import *' is generally unwise, but has been used here to make
it easier for clients to import objects.

* Clean up the fakeswi imports.

The fakeswi imports are now a bit smaller and easier to manage.
Instead of doing conditional things for non-RISC OS, the fakeswi
is used everywhere to ensure that we don't import that SWI interfaces
on RISC OS and have managed interfaces on non-RISC OS systems.